рефакторинг InMemoryFilmStorage

+++ src/main/java/ru/yandex/practicum/filmorate/model/Film.java
@@ -8,13 +8,13 @@ import java.time.LocalDate;
 @Data
 @EqualsAndHashCode
 public class Film {
-    private int id; //целочисленный идентификатор
+    private Long id; //целочисленный идентификатор
     private String name; //название
     private String description; //описание
     private LocalDate releaseDate;//дата релиза
     private long duration; //продолжительность фильма
 
-    public Film(int id, String name, String description, LocalDate releaseDate, long duration) {
+    public Film(long id, String name, String description, LocalDate releaseDate, long duration) {
         this.id = id;
         this.name = name;
         this.description = description;
@@ -22,7 +22,7 @@ public class Film {
         this.duration = duration;
     }
 
-    public int setId(int id) {
+    public long setId(long id) {
         this.id = id;
         return id;
     }
